Generate compilation flavors from .buckconfig file
Summary:
Adds feature where a section of the form cxx#<flavor name> can be
added to the Buck config.  The section is essentially the same
as a normal cxx section, but these settings are only applied
if flavor <flavor name> is used.

Test Plan:
Added unit test AppleCxxPlatformsTest.cxxToolParamsReadFromBuckConfigWithGenFlavor
which tests that the settings of the default Apple platform are still applied
for the generated flavor.