Easy: Remove `@JsonIgnore` on TestResultSummary.getTestCaseName()
Summary:
`ShTest` and a few other places rely on JSON serialization
of `TestResultSummary` objects.

Oddly, `TestResultSummary`'s `getTestCaseName()` was annotated with
`@JsonIgnore`, which meant that field was missing when serialized.
Worse yet, it was set to `null` when deserialized.

There's very little code that handles `TestResultSummary.getTestCaseName()`
returning `null`, so this removes the annotation and updates a test
which relied on the old behavior.

Test Plan: `ant java-test -Dtest.class=EventSerializationTest`