Introduce a Compiler interface for compilers.
Summary:
Different compilers have different flags for the same features. Up until
now we've been able to avoid having an interface for compilers even though
we have one for linkers because clang and gcc are very similar. Sanitising
debug symbols will require passing compiler specific flags, so this adds a
way of doing that and uses it to pass the compilation dir flag to clang.

Test Plan: CI