InputBasedRuleKeyBuilderFactory: use input based rule keys in cache
Summary:
This diff refactors the rule key builder factories a bit so that sub-classes
can make sure the cache news up the the correct type builder key when
calculating the rule key for a `RuleKeyAppendable`.  This fixes a bug in
input-based rule keys, where using the cache pulled in the wrong rule key
builder types.

Since this change meant that input-based and normal rule keys needed
to use different `AppendableRuleKeyCache` types (since the former needed
to cache a tuple of rule key and input list), I ended up merging that
class into the respective factory classes, as it made it easier to force
the `AppendableRuleKeyCache` type to new up a builder of the correct
type.

Test Plan: added unittests