Define flavor `#rn_no_bundle` on `apple_binary` rule to skip bundling RN artifacts.
Summary:
This diff introduces a new flavor on `apple_binary` (and in turn `apple_bundle`) -
`#rn_no_bundle`. Building a target with this flavor will essentially skip
bundling of react native JS and images/assets into the `.app` directory.

The day-to-day workflow of RN developers involves working on JS and assets that
is fetched over the network using the RN development server since it's significantly
faster over re-bundling. With this change, developers can use this flavor (and
manually start the development server) to achieve the same.

Test Plan: Integration test.