[new IJ project gen] Parse java package from source.
Summary:
This introduces a IjProject-specific JavaPackageFinder which
attempts to extract the package from from the actual sources
instead of trusting the convention.
The main reason for this is that IntelliJ will not index
sources that don't have the right package prefix specified
in the project definition.

Test Plan: CI, generate project, manually inspect output.