Generate ZIPs/JARs with deterministic entry ordering
Summary:
Previously, we relied on the filesystem traversal ordering when
constructing ZIPs and JARs, which is filesystem dependennt. This
diff switches to always building with a sorted ordering to make
ZIPs and JARs more deterministic.

As part of this, this diff also made some changes to `ZipStep` to
make it use a `ProjectFilesystem` for all FS operations, so it was
easier to test.

Test Plan:
Added a test to verify that `FakeProjectFilesystem` preserves creation
order, then added a test that verified `ZipStep` produced the same ZIP
regardless of the FS ordering.