Fix IllegalStateException when constructing IjModuleGraph.
Summary:
If TargetNode T1 within module A depends on a TargetNode T2
within a different module and T2 has an exported dependency
of T1 the IjModuleGraph constructor would attempt to make
the module depend on itself and raise an exception.
This change fixes this case by filtering the BuildTargets
coming out of the transitive deps closure resolution.

Test Plan: CI