C/C++: override `PWD` for piped-mode
Summary:
If gcc/clang see `PWD` in the environment, it will use that when
embedding compilation directories into preprocessed sources and
object files *if* it resolves to the same location as `getcwd()`.
Since we expect to see only `getcwd()` when sanitizing the build
in piped mode, this means we can sometimes fail to sanitize the
working directory if the user's `PWD` env var contains symlinks.
This diff fixes this by always overriding `PWD` when we compile
to the project root we use when sanitizing.

Test Plan:
This diff adds more testing around working directory sanitization
handling for both the default platform and the NDK platforms.