CachigBuildEngine: fix re-caching bug with input-based rule keys
Summary:
If we get a cache miss with the main rule key, but get a hit with the
input-based rule key, we re-cache the fetched artifact using the main
rule key, so that subsequent access can get a hit without resorting to
calculating an input-based rule key.

This diff fixes a bug with this behavior, where the build info recorder
would generate empty ZIPs for re-caching, as an input-based rule key
hit never called into `getBuildSteps` which is normally responsible for
recording cached artifacts.  Now, when we get a cache hit, the build
engine records the fetched artifacts back to the recorder, so potential
subsequent re-caching can work.

Test Plan: extended unittest for this specific case