Do not hash `java_library`'s `resources_root` for the target hash
Summary:
If you specified `'.'` as your `resources_root`, the target hashing
code would hash the contents of your directory, which wasn't the
intent of target hashing for this rule.

This change adds a new type of `Hint` on constructor argument
fields that allows you to indicate that the field shouldn't be
used as an input for the rule.

Test Plan: Ran my new test in IntelliJ before (failed) and after (passed)