Fix robolectric res_directories paths.
Summary:
The RobolectricTest rule was using PathSourcePath.toString to construct
the paths list that got passed to the runner. The way the string was
constructed ended up depending on whether the ProjectFileSystem rootPath
was terminated with a slash or not.
The tests didn't catch this because they were using a FakeProjectFilesystem
with an empty root path.
The fix is to have RobolectricTest use the resolver to convert a SourcePath
to a Path and then a String.

Test Plan: build buck, run robolectric tests, make sure they pass.