DX RegisterSpec optimization for de-interning: mapSourceRegisters
Summary:
[Forward Port of 160a87fc81adcc091836b15c54c380069bdf90c5]
An upcoming change will make the RegisterSpec class not interned as a
part of making DX in-process safe.  PhiInsn.mapSourceRegisters is
currently doing an "==" check to determine whether to call
SsaMethod.onSourceChanged.  Without interning, that check might fail
where it would have previously succeeded (for two equivalent but not
reference-equal RegisterSpecs), resulting in an unnecessary call to
onSourceChanged.  (I don't know if this can *actually* happen.)

This diff optimizes that case by doing an early exit if the new and old
source have the same register.  Inspection of the method shows that when
this is the case, it just removes the instruction from a list, then adds
it right back.

This diff should be correct and safe on its own, but it seems unlikely
to have any benefit as long as RegisterSpec is interned.

Test Plan: ant dx-tests