Include DX's classes in buck
Summary:
This is a prerequisite for running DX in-process.

- Add dx/src as a source folder in IntelliJ.
- Add dx/src as a source folder in Eclipse.
- Add dx as a dependency of compile in ant, and add dx.classes.dir to
  the compile classpath.
- Add dx_classes to the runtime classpath.
- Add a monolithic buck module for dx, which is used for buck tests and
  "buck build buck".

This is a little gross, but I couldn't come up with anything cleaner
that would work for "buck test --all".

In build.xml, I couldn't add all of the dx sources to the main javac
step because that is built with -Werror, and dx has warnings.

[Forward Port of 65f51e86c318abb0cd354f884ca8f36f4a8d405b]

Test Plan: Built with in-process DX.