Make @classpath macro output quoted classpath for consumption by javac
Summary:
`javac` reads files provided to it with `@` with a `StreamTokenizer`
configured to treat `#` as a comment character [1]. Buck can generate jar
files with `#` in their name (extracted jars from prebuilt aars are an example).
If such files are on a classpath that is passed to `javac` in a `genrule` with
`@classpath`, `javac` will not parse the classpath as intended.

This diff puts quotes around the entire classpath file so that `javac` knows not to
treat the `#` specially. If `@classpath` is used in genrules that run executables other
than `javac`, those executables will need to gracefully handle the presence of quotes
around the classpath.

[1] http://hg.openjdk.java.net/jdk7/jdk7/langtools/file/ce654f4ecfd8/src/share/classes/com/sun/tools/javac/main/CommandLine.java#l82

Test Plan:
* New unit test
* Create a `genrule` that runs `javac` referencing a classpath containing an `aar`, to compile code that uses classes that appear after the `aar` in the classpath. Verify that the genrule fails before this change, and succeeds after.