Add new `CommandTool` class
Summary:
This diff adds a new `CommandTool` class implementing the `Tool`
abstraction, which aids in composing command-line arguments which
contain references to inputs via `SourcePath`s.  The plan is to
have `BinaryBuildRule` use this instead of a raw list of strings
to support input-based rule keys for `Genrule`s.

Test Plan: added unittests