Change `BinaryBuildRule` to use `Tool` abstraction
Summary:
This diff changes `BinaryBuildRule` to return `Tool` rather than
the raw command formatted as a list of strings.  This allows the
`Tool` to propagate the correct set if inputs that running the command
requires (e.g. if the exectuble is a dynamically linked C/C++ binary
that requires all of it's shared library deps on disk to run) and
should allow using input-based rule keys with `genrule`s.

Test Plan: added additional unittest for execution macro expander