Replace build failure console log "Not all rules succeeded." with a ConsoleEvent and post it to buck event bus.
Summary:
Replace the console log "Not all rules succeeded" for build failure with a
Console Event and post it to buck event bus, so that a ConsoleEventBusListener
can subscribe to the event and print the information as appropriate. This will
make it more consistent with how other real-time build information is displayed.

Test Plan:
Manual Test
1. Modify target source code to introduce a syntax error.
2. Run buck build <target> --keep-going
3. The build failed instead of logging "BUILD FAIL: Not all rules succeeded" to the console directly. The console event is logged by com.facebook.buck.event.listener.LoggingBuildListener.