C/C++ compile: don't hang build on internal errors
Summary:
In rare situations when something has triggered an internal error
to be thrown from inside the C/C++ step, which doesn't kill the running
process, it's possible for the step to deadlock.  This happens if
an error is triggered before we start and finish reading all the
stdout coming from the preprocess step, in which case the process
itself hangs trying to write stdout and Buck gets stuck trying to
finish reading from stderr inside the `close` method of the corresponding
`FunctionLineProcessorThread`.

This diff moves the blocking done in `ManagedThread` to be explicit,
then creates a sub-class in the C/C++ step which destroys the process
when it's being closed, so that the reading thread will return from
native code and be able to exit.

Test Plan:
It's pretty difficult to test this, as it requires triggering an
internal exception which does *not* kill the preprocess/compile step.
The only time we saw this was due to another bug, fixes in another
diff.