CachingBuildEngine: maintain recorded paths in build metadata
Summary:
We only find the paths a `BuildRule` records to cache when building
it.  Normally this was fine, since we only needed to get this list
when uploading to cache, and we used to only do this when we ended
up building locally.  However, now that we have input-based rule keys,
we need to still upload to the cache when we get a local key match or
a cache fetch using input-based rule keys, since this would only happen
if we had originally had a miss fetching via the normal rule key and
we therefore want to re-index the artifact using it's normal rule key
so that subsequent fetches using it result in a hit.

To do this, however, we need the get the recorded path list without
executing the rule.  This diff has the `BuildInfoRecorder` save the
recorded paths in it's build metadata, and has the `CachingBuildEngine`
recover the recorded paths from it when the target wasn't build locally.

Test Plan: added unittests