Reduce object churn in DebugPathSanitizer.sanitize
Summary:
This diff reduces the object churn in the VM in three ways;

- It avoids the creation of and population of an ImmutableBiMap with the same
data as 'other' if no working directory is specified in getAllPaths.
- It caches the values returned by getAllPaths to avoid the multiple
ImmutableBiMap objects being created for the same working directory
- It checks to see if String.replace would result in a change to the string
in sanitize, and if it doesn't it avoids the object creation overhead involved
in calls to replace.

With these changes compilation of Facebook for Android is showing a
significant reduction in memory churn for DebugPathSanitizer.sanitize
(~75% reduction)

Test Plan:
- ant test
- Check build time on  buck build --no-cache native/liger:liger
- Verify memory usage reduction using a profiler