C/C++: add `env` parameter for test rules
Summary:
Adds an `env` parameter, which allows the user to pass in environment
variables which are set when the test runs.

Test Plan:
Add tests to verify that the env var propagates.  It's pretty difficult
to add actual integration tests to verify that running tests have this
set, as we'd need to import the gtest/boost test libs into Buck.