C/C++: merge `ld` and `cxxld` concepts
Summary:
We started off distinguishing between the linker (`ld`) and the
linker driver (`cxxld`).  However, this never became useful in any
way, and makes the code particularly confusing/complex.

This diff merges these two concepts into a single one to represent
the linker (previously the linker driver) and it's associated flags.
The main UI difference is the lack of `cxxld` and `cxxldflags` and
that now `ldflags` go the linker driver (and may need to be prefixed
with `-Wl,`).

Test Plan: `buck test --all`