Expand pragmas after preprocessing.
Summary:
Apple clang is normally invoked on source code without preprocessing it first.
Because of that it has some issues when called separately for preprocessing and
compiling, specifically around how it handles pragmas. To work around that, we
disable pragma expansion in the preprocessor by renaming the `_Pragma` built-in,
and do it with the line processor after the preprocessing.

Test Plan: CI