Import bazel's query engine to buck.
Summary:
Import bazel's query engine code to third-party/.

The only difference in the imported "query2.engine" package is that I did not include the class BlazeQueryEvalResult. The reason is that this class has dependencies to a class outside the package and it won't be used for buck query.

Added an exception to the linter to ignore bazel code.

Test Plan: This diff just imports the bazel code and it does not interact with buck for now.