Stop the build on the first test failure
Summary:
Finding out why builds have failed from the CI server is incredibly
painful because ant runs every test and outputs all the results, so
a human has to read everything. It's nicer to people to stop an ant
build as soon as a test failure is detected.

The buck test invocation we do will still run every test.

Test Plan:
Introduce a test failure, run "ant test", so that not all tests are
run.