Add file existence cache to FilesystemBackedBuildFileTree
Summary:
In his commit `don't do an eager filesystem crawl to create the build file tree`,
@natthu removed the state from `BuildFileTree` and split it into two classes
(`InMemoryBuildFileTree` and `FilesystemBackedBuildFileTree`).

The new class `FilesystemBackedBuildFileTree` always talks to the filesystem
to check for file existence (mainly from `TargetNode` and `Parser`).

`Parser` already had code to throw away and create a new
`BuildFileTree` between commands, so it's safe to add state to
this object to improve performance.

In this diff, I add a simple `LoadingCache` to
`FilesystemBackedBuildFileTree` so `TargetNode`'s constructor doesn't
repeatedly hit the filesystem tens of thousands of times for targets
with large deps / srcs when validating the sources.

Test Plan:
`buck build [target]` in repository with React Native target
with tens of thousands of targets.

Previously, creating the target and action graphs took about 74 seconds.
Now, it takes around 59 seconds.