Add more Xcode file type identifiers.
Summary:
Loading a project in Xcode scales quadratically with the number of
files it has to check for the type. This adds all the types defined
in Xcode 7 beta 4 to make sure the number of files without an
explicit file type is as small as possible.

Test Plan:
CI
open project with many files with extensions that weren't in the
list before, notice performance improvement