Allow unit tests to access anything in the bootstrapper
Summary:
Previously `DEFS` was set up to allow unit tests to access
the event API, because in a unit test context there was no bootstrapper
to load it. However, really the right fix there was to allow unit tests
to access anything in the bootstrapper. This commit takes care of that.

Test Plan:
The motivation for this change was that tests were failing when we added
new classes to the bootstrapper in subsequent commits. Therefore, the
test plan is to make sure all of those tests now pass.