Add an OSO sanitiser.
Summary:
Apple's linker adds STAB OSO entries that refer to the object files it linked
when the object files contain debug symbols. These entries contain
absolute paths as names and time stamps as values, so they need to be
sanitised. This adds a new scrubber that postprocesses the linked file to
change the OSO references to be relative, and sets the timestamps to far
into the future.

Test Plan:
CI
new tests