Support multiple test suites in gtest results
Summary:
The result parser originally limited its search for test results to the
first testsuite found in the results file for gtest. This will look for
all testcase elements.

Test Plan:
Run a buck test against a gtest with multiple test suites.
Ensure that all test results are accounted for