Use a more memory efficient method of line reading
Summary:
The BufferedReader.readLine() method constructs multiple StringBuilder
objects and Strings which causes a noticable memory churn.

This diff replaces the most common calls with a method which has a single
buffer for a given stream and re-uses it. The String creation creates
a copy of that buffer so it's safe for re-use once it has been converted
into a string.