Unsanitise debug compilation dir when running the built application in Xcode.
Summary:
Xcode sets breakpoints using absolute paths, and lldb matches paths without
resolving them. This means the sanitised debug compilation dir will cause
breakpoints to be able to resolve properly. This adds an extra step when
running an application built with Buck from Xcode, where the debug
compilation dir is replaced with the absolute path to the project root,
padded using the hostname, so the absolute path matches.

Test Plan:
CI
generate a project with --build-with-buck, set a breakpoint in Xcode