Support multiple queries in a single buck query call.
Summary:
Many use cases of the buck audit commands involve aggregating results by input targets.
This is the same as doing multiple queries at once.

In order to replace those with buck query, we need to support multiple queries in a single buck call.
This is done by definining a query format and providing a list of input targets. Example:
  buck query "deps(%s)" //example:foo //example:bar

This buck query call will determine the transitive dependencies of the 2 inputs and output:
- a json with the mapping input -> result, if --json is used
- the query results, one per line, otherwise.

Test Plan: Unit tests.