buck.py perf [2/7]: Easy: Add new parser events for perf tracing
Summary:
This is part 2 of 7 of a series of big performance fixes to
the `buck.py` parser of `BUCK` files.

In this diff, I add a number of new `BuckEvent`s to let me monitor
where all the time goes during invocation of `buck.py` and parsing
of its output.

This diff just includes the new `BuckEvent` classes -- the next diff
will actually use them.

Test Plan: CI