FileHashCache: propagate `IOException` on `get`
Summary:
Previously we'd catch `IOException`s within the `DefaultFileHashCache`
and propagate it as a `RuntimeException`.  This diff preserves the
`IOException` type, as it'll be useful when supporting dep files, in
which case we'll need to know when an input from a dep file was deleted
(and trigger a rebuild rather than failing internally).  This diff will
allow us to catch the `IOException` and handle the `FileNotFoundException`
specially.

Test Plan: added unittest for expecting the `FileNotFoundException`.