iOS codesigning and provisioning file support
Summary:
This implements support for generating bundles with valid provisioning profiles and code signatures for device installation, if we're building for a device flavor:

1. Selecting the best mobile provisioning file based on the app identifier.
2. Copying the file
3. Generating an entitlements plist file for code signing.
4. Running `/usr/bin/codesign` to sign the bundle.

The bundle can be manually made into an .ipa and installed by hand onto a device.

Test Plan:
1. `ant test`

2. Make sure you have a generic development provisioning profile installed valid for your app's identifier as well as a code signing identity in your keychain as per Apple's developer instructions, then build an app with a device flavor.

```
buck build //Apps/MyApp:MyApp#iphoneos-arm64
```

See there's a `.mobileprovision` file in the output directory, as well as a `_CodeSignature` directory.