buck.py perf [7/7]: Add optional BSER buck.py output format support
Summary:
This is part 7 of 7 of a series of big performance fixes to
the `buck.py` parser of `BUCK` files.

In this diff, I add optional support for the BSER binary encoding to
`buck.py` including a Java BSER decoder. This format is documented
here:

https://facebook.github.io/watchman/docs/bser.html

I implemented the entire format except for the `TEMPLATE` data
type, since I don't need to use it.

This results in about a 20% improvement in performance of parsing
`BUCK` files with lots of globs.

Test Plan:
Unit tests included. `ant java-test -Dtest.class=BserDeserializerTest`

`buck build largeproject` with `watchman` 3.5.0 in `$PATH`.
Confirmed project took about 60 seconds to parse `BUCK` files. in
Chrome trace, confirmed that invoking `buck.py` was the bulk of the time.

Updated `$PATH` to point to `watchman` 3.6.0. Re-ran test.
Before, it took about 60 seconds. Now, it takes about 20 seconds.