CachingBuildEngine: support dependency-file rule keys
Summary:
This diff adds dep-file rule key support to the build engine.
For rules that support them, the build engine gets from them a
list of explicitly used inputs and records them (locally and to
the cache), along with a dependency file rule key, for next time.

On subsequent runs, if the dep-file and associated rule keys
are availabl from the last run, a new dependency file rule key
can be calclulated to check if a rebuild can be avoided.

Addresses part of #298

Test Plan: added unittests