Bump to LLVM 3.5
Summary:
This diff bumps the default clang version to 3.5 to ensure compatibility with
recent NDK distributions.

The previous default (3.4) was removed in the most recent NDK (10e) and so
using 3.4 as the default makes buck fail when it's used with the latest NDK.

3.5 exists in NDK 10c and above, so using 3.5 maintains compatibility with
more versions than going directly to 3.6 (which currently is only in 10e).

3.5 does introduce warnings for a couple of flags we were using by default,
and this diff removes those switches because it causes the tests to fail
because they treat warnings as errors.

Test Plan: - ant test