Fix Watchman globbing when watch created for parent of Buck project root
Summary:
My previous Watchman glob logic issued queries assuming that the
Buck project root was the same as the Watchman watch root.

In his diff `Add project awareness to buck watchman interaction`,
@DurhamG added support for running Buck with a Watchman watch for a
parent directory of the Buck project root. For example, the watch
might be `/path/to/watch` and the Buck project might be in a subdir
`/path/to/watch/project`.

When using such a project layout, my new Watchman glob would fail when
incorrectly trying to query from the root of the project instead of
the watch root:

  % buck build //path/to:lib
  Using buckd.
  pywatchman.CommandError: watchman command error: unable to resolve root /path/to/watch/project: directory /path/to/watch/project is not watched

This fixes the bug by passing the optional (watch root, project
subdir) data down to `buck.py` and ensuring all queries are issued to
the watch root.

We were already using the `relative_root` feature to limit results to
a subdirectory of the project, so I just made sure to prepend
the project subdirectory to the relative root.

Test Plan:
Created `.buckconfig.local` with contents:

  [project]
    glob_handler = watchman

Ran `buck build buck`, confirmed project prefix was absent in logs,
Watchman globbing was enabled, and Watchman globs worked.

Repeated in a project with a watch root outside the project root, confirmed
project prefix was present in logs, Watchman globbing was enabled, and
Watchman globs worked.

Removed `.buckconfig.local`, repeated tests, confirmed Watchman globbing
was disabled and Python globs worked.