Fix AndroidBinaryIntegrationTest with python3 as default
Summary:
On systems on which python refers to python3, the
AndroidBinaryIntegrationTest fails in the testPreprocessorForcesReDex
test, because of a python script that is not compatible with python3.
PEP 0394 suggests that python2 should always refer to a valid python2
interpreter.  Adjust the shebang line to make the script use python2,
which fixes the test.

Closes https://github.com/facebook/buck/pull/389

Test plan: buck test

Test Plan: Imported from GitHub, without a `Test Plan:` line.