Integration test to expose bug in android_react_native_library rule.
Summary:
`AndroidReactNativeLibrary` implements the `AndroidPackageable` interface, in
order to provide the JS bundle as an asset. However, it doesn't own the JS bundle
(it gets the path to the assets dir from `ReactNativeBundle`). Now,
`AaptPackageResources` directly depends only on `AndroidReactNativeLibrary`, as
a result, in top-down building, when its rule key changes, it only gets outputs
from `AndroidReactNativeLibrary`, which does not cache anything, and so the
resulting APK does not contain the JS bundle.

I plan to fix this in a subsequent diff.

Test Plan: Integration test.