Easy: Lower log level when buck.py parser exits unexpectedly
Summary:
When parsing a `BUCK` file, if the python process exits
unexpectedly (say, due to a syntax error), the `BSER` parser will
throw `BserDeserializer.BserEofException`.

In this case, we already have user-visible messaging to include
the Python stack trace and let them know the build failed.

Before this diff, we use `LOG.error()` to include the `BserEofException`
stack trace and print it to the console. However, the Java stack
trace isn't really useful to the user here.

Since we already show the Python stack trace and let the user know
the build failed, this diff simply lowers the log level from `LOG.error()`
to `LOG.warn()` (which prints the Java stack trace to the log file but
not to console).

Test Plan:
`buck audit dependencies //path/to:lib` where `BUCK` file
contained a call to an undefined function