Support reverse dependencies function in buck query.
Summary:
Added a new operator - `rdeps`. Given targets U and T, `rdeps` finds
the  reverse dependencies of T in the transitive closure of U (the `universe`).

There is an optional third argument which represents the depth of search for
the reverse dependencies.

Updated the `buck query` doc to describe this operator.

Test Plan: integration tests.