Remove extra copy of getIgnorePaths()
Summary:
In his commit `Decouple BuckConfig from ProjectFilesystem`,
@shs96c made a copy of the underlying logic of `BuckConfig.getIgnorePaths()`
but didn't get to delete the original copy.

This removes the original copy and moves the tests over to `ProjectFilesystemTest`.

When I did that, I found the utility to create a `Config` from a string
was not exported from `ConfigTest`, so I moved it into a new test utility
class `ConfigBuilder`.

Test Plan:
  ant java-test -Dtest.class=ProjectFilesystemTest
  buck test //test/com/facebook/buck/io:io