Register debug symbols after extracting them.
Summary:
Now that the debug symbols are not in the app bundle lldb will not find them
automatically when they're loaded in the simulator. This loads the debug
symbols in lldb after extracting them so that lldb will remember them the
next time it loads a binary with the same UUID.

Test Plan:
CI
Build and run an app with Buck, verify debugging works.