Use `BuildTargetSourcePath`s to fix missing dependency on `ReactNativeBundle`.
Summary:
Earlier, I made `AndroidReactNativeLibrary` implement `HasRuntimeDeps` to
essentially "export" the `ReactNativeBundle` dependency. As @coneko pointed out,
this was not the right fix. @andrewjcg had mentioned on that diff that using
`BuildTargetSourcePath` should take care of it, but it did not work when I tried
using it. Turns out `AndroidBinaryGraphEnhancer` would not filter out build
targets from the `assetsDirectory`.

This diff also makes `ReactNativeLibraryGraphEnhancer` use a `BuildTargetSourcePath`
so that rules that depend on the resource directory of the graph enhanced
`AndroidResource` rule also depend on the `ReactNativeBundle`.

Test Plan: AndroidReactNativeIntegrationTest