Add a labels operator to buck query.
Summary:
Adds a new `labels` operator to `buck query`. This operator provides a way to query the content of attributes of build rules. For example,

  $ buck query "labels('srcs', '//foo:bar')"

evaluates to the files/ build targets in the `srcs` atribute of the target `//foo:bar`.

The attribute content is found via traversing the ConstructorArg via type coercers.
Only build targets and file paths are retrieved.

Test Plan: Integration tests + running on the shell.