DefaultJavaLibrary: compute transitive classpath lazily
Summary:
The transitive classpath collection process is extremely costly,
and we only need it for suggesting deps when a build error occurs.
This diff moves it's generation into the supplier we call when we
see an error to avoid this cost in the common case.

Test Plan:
Profiled local builds, which spent less time collectin the transitive
classpath map.