AbstractBuildRule: keep dependencies as suppliers
Summary:
We already construct the params using `Suppliers`, but we evaluate
them as soon as we construct the `AbstractBuildRule`.  This diff
keeps the dep collections as suppliers, and waits to evaluate them
when `getDeps` is called.

This will allow us to defer some of the work we were previously doing
during the serial action graph construction phase to a more parallel
phase.

Test Plan: CI