Correctly installs exopackage builds with libraries packaged as assets
Summary: Buck outputs cxx libraries in different folders depending on whether they are marked as can_be_asset, but only looks in the folder for non-asset libraries when running exopackage. This diff fixes that by adding which folder each cxx library is in into the metadata and parsing it accordingly.

Test Plan: Mark many cxx libraries as can_be_asset and verify that buck install facebook for android works.