Simplify how we deal with resources inside an `AndroidPrebuiltAar`.
Summary:
http://tools.android.com/tech-docs/new-build-system/aar-format
^ mentions that `R.txt` is mandatory inside an AAR. So, instead of creating
an `AndroidResouce` build rule to process the resources, which essentially
just produces the same `R.txt`, we can use the packaged file directly.
However, we also need to make sure that `AndroidPrebuiltAar` provides all the
necessary artifacts to the `AndroidPackagebleCollector`.

Also, although the format says `R.txt` is mandatory, I came across a couple
popular AARs on the internet which did not have it. I believe the reason is
that they do not contain/refer to any resources, hence no need for `R.txt`.
In order to support such AARs, I have added a `TouchStep` to make sure that
the unzip dir always contains `R.txt`.

Test Plan: CI