DefaultJavaLibrary: support input-based rule keys
Summary:
This diff switches `DefaultJavaLibrary` rules to use input-based
rule keys, instead of ABI rule keys, so that:
1) We avoid rebuilds if generated sources change.
2) We get cache hits off the input-based rule key.

As with ABI rule keys, input-based rule keys hash the contents of
the ABI JARs of dependencies.  Since input-based rule keys are a more
generalized form of ABI rule keys, this diff also spins out the ABI
calculation of each Java library into it's own rule, which we wrap
in `SourcePath`s for depedents to add to their rule keys.  This means
when calculating the input-based rule key, we'll only consider the
contents of the ABI JAR, as desired.

Test Plan: Added and updated tests.