Build RuleKeyFactories in the CachingBuildEngine.
Summary:
Dependency Injection is all well and good, but copy & paste
is widely acknowledged to be a less than ideal way to write
code. Reduce copy-pasta.

This also makes it possible to decorate the file hash cache
used by the CachingBuildEngine on a per-repo basis which is
the actual reason for this diff.

Test Plan: buck test --all