New class SimpleProcessListener and components to handle encoding/decoding bytes to Unicode
Summary:
The `ListeningProcessExecutor` class streams stdio bytes to
and from a listener.

Much of the time, we need to encode and decode these bytes to and from
Unicode. I added streaming support to the third-party `NuProcess`
library to efficiently handle this.

This diff wraps the `NuProcess` encoding/decoding support in a few new
classes:

1) `ListeningCharsetEncoder`: Encodes bytes to be sent to stdin to Unicode
and streams them to the process

2) `ListeningCharsetDecoder`: Decodes bytes sent to stdout/stderr to Unicode
and streams them to the listener

3) `AbstractCharsetProcessListener`: Wraps a `ListeningCharsetEncoder` for
stdin and two `ListeningCharsetDecoder`s for stdout and stderr and streams
them to the listener

4) `SimpleProcessListener`: Subclass of
`AbstractCharsetProcessListener` which takes an optional string to
write to stdin and handles streaming it to stdin. Buffers all decoded
string data sent to stdout and stderr in memory.

I also included a fix for "soft close" of stdin to fix a few race conditions
in `NuProcess`. I sent this upstream:

https://github.com/brettwooldridge/NuProcess/pull/42

Test Plan: Unit test included. `ant java-test -Dtest.class SimpleProcessListener`