buck test: support delegating to an external test runner
Summary:
This adds the ground work to supporting using external test runners
with Buck.  The basic approach is to provide a `.buckconfig` setting
for the user to specify a command to run in lieu of Buck's internal
command runner.  If found we build all the tests and call into the
external test runner passing it "info" files describing how to run each
test rule.

The info file format is described in `AbstractExternalTestRunnerTestSpec`.
Tests can opt-in to supporting external test running by implementing
the `ExternalTestRunnerRule` interface and returning the above object.

Test Plan:
I've added support to `sh_test` and added a few tests for pass and
failure using a dead simple external test runner.