New script scripts/create_tag.sh
Summary:
To support building Buck outside the git repo (for external releases), we need a metadata file in the Github repo from which the build script can read a unique identifier for the release (to use instead of the git commit hash).

This adds a new script `scripts/create_tag.sh` which creates and commits a `.buckrelease` file in the root of the repo containing a
version in the form `vYYYY.MM.DD.NN` (where `NN` is an integer identifying the Nth release of the day).

It then tags the release and optionally pushes the `.buckrelease` and tag to Github.

Test Plan: `./scripts/create_tag.sh`