New config option to forbid symlinks in source file paths
Summary:
Currently, Buck's parser cache works on the assumption
that a change to a path foo/bar/baz.txt can only invalidate build
targets defined in that directory or any of its parents:

  foo/bar/BUCK
  foo/BUCK
  BUCK

Symbolic links break this assumption; if foo/bar is actually a symbolic link to baz/blech,
then we also have to invalidate the build targets defined in baz/blech/BUCK, and baz/BUCK.

This turns out to be pretty hard to implement correctly, so Buck's parser cache
conservatively falls back and dumps a huge section of the cache whenever it sees a change to
any files whose paths contain a symbolic link as a path component.

Even though we sternly warn when we do this, the net effect is that
users complain that Buck is slow.

This diff introduces a new Buck config option:

  [project]
    allow_symlinks = forbid

which changes the warning to an exception explaining what's going on.

When I fixed this, the test I added revealed that the logic in
`Parser` to build the map of (symlink -> target) pairs (only used in
the debug log message up until now) didn't actually work. I fixed this
bug to ensure the test passed.

Test Plan: Integration test included. `ant java-test -Dtest.class=ParserTest`