Stop building pywatchman via ant
Summary:
Building pywatchman by running
`third-party/py/pywatchman/setup.py` in the ant build has caused us no
end of trouble.

Since I replaced the native BSER module in pywatchman with a pure
Python implementation, we no longer need to run setup.py at all.

This removes the ant build of pywatchman in favor of always using
the Buck resource (which we put in place for the PEX build).

Test Plan:
Tested with and without watchman glob from both repo build
and PEX package build by running:

  ant && touch build/successful-build
  ./bin/buck build buck
  rm -rf buck-out/bin buck-out/gen/src
  ./buck-out/gen/programs/buck.pex build buck

Confirmed in both logs that with watchman glob enabled, pywatchman was
enabled and BSER encoding was used, and without it pywatchman was
not enabled and JSON encoding was used.