Added CompilerErrorEvent and made WebServerBuckEventListener listen for it.
Summary:
Added CompilerErrorEvent to allow web socket clients to get the error message. The ideabuck plugin will use this to show in the interface a clickable link to the file with the error.

This works only for java builds (since the event is generated only when JavacStep fails.

Test Plan:
Try to buck build on "broken" sources (eg, deliberately use a non existent variable, etc). The console will contain an error message about the missing symbol. The websocket server will report a CompilerError event, that will contain the error message, similar to:

  Message: {"timestamp":...,"nanoTime":...,"threadId":...,"buildId":"...","eventKey":{"value":...},"target":"//java/com/...","error":"error will go here","suggestions":[],"compilerType":"Java","type":"CompilerErrorEvent"}